verbs: Consolidate duplicate code in create_rwq_ind_table
The usual pattern for an _ex API with a core flex array is to have the
core code allocate the variable length command buffer and memcpy the udata
into the end of it. Tidy this up to follow the usual pattern.

Since no drivers have udata here don't bother passing anything from the
driver.

Also update the PABI to v22.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>