ibacm: Allocate end-point addresses dynamically in provider - v2
ibacm's default provider has a fixed number (4) of end-point
addresses. An endpoint is associated with the node GUID. Hence, if a
system is brought up with VFs enabled, the default provider will not
be able to handle the end-point addresses.

acmp uses the address of a particular addr_info entry in struct
acmp_ep as its addr_context, stored ibacm's data structures. However,
since these addresses now may change, due to realloc(), we change the
addr_context to a small struct containing a pointer to the end-point
and an addr_inx.

This way, we can safely get the correct address of a realloc'd
addr_info entry.

Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>