ibacm: Unable to resurrect an interface
When an IB port has been brought back to Active state, after being
down, ibacm gets an event about it. It will then (re) enumerate the
devices, and does so by executing an ioctl with SIOCGIFCONF. This
particular ioctl will only return interfaces that are "running".

There may be a delay after the IB port becomes Active until its
address has been provisioned, and becomes "running". If ibacm attempts
to associate IPoIB interfaces to the port during this interval, it
will not see the interface because it is not "running".

Later, when ibacm is asked for a Path Record (PR) using the IP address
of the resurrected IPoIB interface, it will not be able to find the
associated end-point (EP), and the following is printed in the log:

acm_svr_resolve_path: notice - unknown local end point address

The bug can be provoked by the following script. We have a single HCA
with two ports, the IPoIB interfaces are named stib{0,1}, the IP
address of the first interface is 192.168.200.200, and the remote IP
address is 192.168.200.202. The LID of the IB switch is 1 and the
switch port number connected to port 1 of the HCA is 22.

<script>
 #!/bin/bash

ibportstate -P 2 1 22 disable

 # move the IP address
ip addr del 192.168.200.200/24 dev stib0
ip addr add 192.168.200.200/24 dev stib1

ibportstate -P 2 1 22 enable

 # Wait until port becomes active again
while [[ $(ibstat|grep State:|grep -c Active) != 2 ]]; do
   echo -n "."; sleep 1
done
echo

 # give ibacm time to re-enumerate the interfaces
sleep 1

 # move the IP address back
ip addr del 192.168.200.200/24 dev stib1
ip addr add 192.168.200.200/24 dev stib0

 # take the other port down, so we are sure we use stib0
ip link set down dev stib1

 # Start a utility requesting a PR
qperf 192.168.200.202 -cm1 rc_bw

 # check for failure
grep "unknown local end point" ibacm.log

 # restore the state
ip link set up dev stib1
</script>

This fix depends on the commit that re-factors the use of ioctl in
acm_if_iter(), and instead uses netlink. Now, by reducing the
requirements of the state of the interface, the EP is added, and
afterwards, when an address is assigned, it is associated with the EP.

This commit is a new implementation of
https://patchwork.kernel.org/patch/10748357, which was NAKed.

Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>

v1 -> v2:
   * Removed Gerrit's Change-Id tag (Håkon)